Update ExplainAdminCommand.php
It's quite possible that your application requires the tokenStorage in order to decide what filters to show. If you do this, it means that the AuthorizationChecker.php throws an exception due to `isGranted` check requiring a set token. This is unavoidable, and prevents the command from continuing, however, it would be better to just continue gracefully. The command doesn't change anything so a try catch allows the getForm reflection debugging to execution plus the other remaining commands.